rename state_machines to nb_device
